# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gfact API",
    "version": "1.0.0",
    "description": "API de gerenciamento de subscrições do Gfact"
  },
  "servers": [
    {
      "url": "http://localhost:4002/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "JsonSchema": {
        "type": "object",
        "properties": {
          "Identificacao do Emitente": {
            "type": "object",
            "properties": {
              "Nome": {
                "type": "string"
              },
              "Morada": {
                "type": "string"
              },
              "NIF": {
                "type": "string"
              },
              "Contacto": {
                "type": "string"
              }
            }
          },
          "Identificacao do Cliente": {
            "type": "object",
            "properties": {
              "Nome": {
                "type": "string"
              },
              "NIF": {
                "type": "string"
              },
              "Morada": {
                "type": "string"
              }
            }
          },
          "Data e Hora da Factura": {
            "type": "object",
            "properties": {
              "Data": {
                "type": "string"
              },
              "Hora": {
                "type": "string"
              }
            }
          },
          "Numero da Factura": {
            "type": "object",
            "properties": {
              "Factura-recibo": {
                "type": "string"
              },
              "Documentos referenciados": {
                "type": "string"
              }
            }
          },
          "Valor Total": {
            "type": "string",
            "description": "Exemplo \"KZ 500,00\""
          },
          "IVA e Base Tributavel": {
            "type": "object",
            "properties": {
              "Base tributavel": {
                "type": "string"
              },
              "IVA (%)": {
                "type": "string"
              },
              "Valor Total com IVA": {
                "type": "string"
              }
            }
          },
          "Pagamento": {
            "type": "object",
            "properties": {
              "Forma de pagamento": {
                "type": "string"
              },
              "Valor": {
                "type": "string"
              }
            }
          },
          "Outras Informacoes": {
            "type": "object",
            "properties": {
              "Software": {
                "type": "string"
              },
              "Emp.": {
                "type": "string"
              },
              "Data de processamento": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "icon": {
            "type": "string"
          },
          "colorHex": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deleted": {
            "type": "boolean"
          },
          "deletedBy": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Type": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "jsonData": {
            "type": "object",
            "$ref": "#/components/schemas/JsonSchema"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "categoryId": {
            "type": "string"
          },
          "typeId": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "type": {
            "$ref": "#/components/schemas/Type"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/process-invoice": {
      "post": {
        "summary": "Processa uma nova fatura",
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fatura processada e salva com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/JsonSchema"
                }
              }
            }
          },
          "400": {
            "description": "Nenhum arquivo enviado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Nenhum arquivo enviado"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Limite de faturas atingido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Limite de faturas atingido"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invoices": {
      "get": {
        "summary": "Lista as faturas do usuário autenticado",
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "typeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de faturas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{id}": {
      "get": {
        "summary": "Busca uma fatura específica pelo ID",
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fatura encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Fatura não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não encontrado"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleta uma fatura específica",
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fatura excluída com sucesso"
          },
          "401": {
            "description": "Usuário não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Fatura não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Fatura não encontrada"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invoices/all": {
      "get": {
        "summary": "Lista todas as faturas (admin)",
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "typeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista completa de faturas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}